dashboard/backend/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
pandas==2.1.0
numpy==1.25.2
scikit-learn==1.3.0
joblib==1.3.2